requirements.txt
pandas==2.2.1
chardet==5.2.0
spacy==3.7.4
Telethon==1.34.0
config==0.5.1
python-decouple==3.8
sentence-transformers==2.6.1
transformers==4.39.3